pyproject.toml
[project]
name = "two-d-library"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
]

[tool.mypy]
strict = true
# Disallow dynamic typing: use of the dynamic 'Any' type under certain conditions
disallow_any_unimported = true
# None and Optional handling: adjust how values of type 'None' are handled
no_implicit_optional = true
# Configuring warnings: detect code that is sound but redundant or problematic
warn_unreachable = true
# Configuring error messages
show_error_codes = true
pretty = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Rules that conflict with Ruff formatter:
    "COM812",  # Trailing comma missing
    # Other:
    "D105",  # Missing docstring in magic method
    "S311",  # Standard pseudo-random generators are not suitable for cryptographic purposes

]

[tool.ruff.lint.per-file-ignores]
# Not relevant in tests:
"tests/*" = [
    "S101",  # Use of assert detected
    "PLR2004",  # Magic value used in comparison
]

[tool.ruff.lint.pydocstyle]
convention = "google"
